requirements.txt
# Core Trading Framework
lumibot>=2.8.0

# Broker Integration
alpaca-py>=0.9.0
alpaca-trade-api>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.15.0
statsmodels>=0.14.0

# Market Data
yfinance>=0.2.28

# RiskLabAI and Dependencies
RiskLabAI>=1.0.7
memory-profiler>=0.61.0
sympy>=1.12
numba>=0.58.0
ta>=0.11.0
quantecon>=0.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0
requests>=2.31.0
coloredlogs>=15.0

#Technical Indicators (Optional: requires binary installation)
ta-lib>=0.4.0

holidays

# AWS Cloud infrastructure
boto3>=1.28.0
